+++ omh_shim/_schema_loader.py
@@ -30,6 +30,7 @@ _FILENAMES: dict[str, str] = {
     "ieee:sleep-stage-summary:1.0": "data/ieee_sleep-stage-summary_1-0.json",
     "ieee:physical-activity:1.0": "data/ieee_physical-activity_1-0.json",
     "ieee:sleep-episode:1.0": "data/ieee_sleep-episode_1-0.json",
+    "ieee:total-sleep-time:1.0": "data/ieee_total-sleep-time_1-0.json",
     "ieee:header:1.0": "metadata/header-1.0.json",
 }